番組詳細で overlap 対応 #102

+++ src/client/Component/Program/ProgramInfoActionComponent.ts
@@ -64,9 +64,11 @@ class ProgramInfoActionComponent extends Component<void> {
         } else if (reserve.status === 'reserve' || reserve.status === 'conflict') {
             name = '削除';
             onclick = () => { return this.viewModel.deleteReserve(); };
-        } else {
+        } else if (reserve.status === 'skip') {
             name = '除外解除';
             onclick = () => { return this.viewModel.deleteSkip(); };
+        } else {
+            return null;
         }
 
         return m('button', {
